Development

+++ coursework/src/main/java/by/grsu/coursework/Controllers/BlogController.java
@@ -2,6 +2,7 @@ package by.grsu.coursework.Controllers;
 
 import by.grsu.coursework.models.Post;
 import by.grsu.coursework.repo.PostRepository;
+import by.grsu.coursework.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,19 +24,20 @@ public class BlogController {
     public String blogMain(Model model) {
         Iterable<Post> posts = postRepository.findAll();
         model.addAttribute("posts", posts);
-        model.addAttribute("title", "Блог");
-        return "blog-main";
+        model.addAttribute("title", "Forum");
+        return "blog";
     }
 
     @GetMapping("/blog/add")
     public String blogAdd(Model model) {
-        model.addAttribute("title", "Добавление статьи");
+        model.addAttribute("title", "New article");
         return "blog-add";
     }
 
     @PostMapping("/blog/add")
-    public String blogPostAdd(@RequestParam String title, @RequestParam String anons, @RequestParam String full_text, Model model) {
-        Post post = new Post(title, anons, full_text);
+    public String blogPostAdd(@RequestParam String title, @RequestParam String anons, @RequestParam String full_text, @RequestParam String author, Model model) {
+        author = User.getCurrentUsername();
+        Post post = new Post(title, anons, full_text, author);
         postRepository.save(post);
         return "redirect:/blog";
     }
@@ -46,11 +48,15 @@ public class BlogController {
             return "redirect:/blog";
         }
 
+        Post postData = postRepository.findById(id).orElseThrow();
+        postData.setViews(postData.getViews() + 1);
+        postRepository.save(postData);
+
         Optional<Post> post = postRepository.findById(id);
         ArrayList<Post> res = new ArrayList<>();
         post.ifPresent(res::add);
         model.addAttribute("post", res);
-        return "blog-details";
+        return "post-details";
     }
 
     @GetMapping("/blog/{id}/edit")
@@ -67,12 +73,18 @@ public class BlogController {
     }
 
     @PostMapping("/blog/{id}/edit")
-    public String blogPostUpdate(@PathVariable(value = "id") long id, @RequestParam String title, @RequestParam String anons, @RequestParam String full_text, Model model) {
+    public String blogPostUpdate(@PathVariable(value = "id") long id, @RequestParam String title, @RequestParam String anons, @RequestParam String full_text, @RequestParam String author, Model model) {
         Post post = postRepository.findById(id).orElseThrow();
-        post.setTitle(title);
-        post.setAnons(anons);
-        post.setFull_text(full_text);
-        postRepository.save(post);
+
+        if (User.getCurrentUsername().equals(post.getAuthor())) {
+            author = User.getCurrentUsername();
+            post.setTitle(title);
+            post.setAnons(anons);
+            post.setFull_text(full_text);
+            post.setAuthor(author);
+            postRepository.save(post);
+        }
+        else return "redirect:/accessDenied";
 
         return "redirect:/blog";
     }
@@ -80,9 +92,13 @@ public class BlogController {
     @PostMapping("/blog/{id}/remove")
     public String blogPostDelete(@PathVariable(value = "id") long id, Model model) {
         Post post = postRepository.findById(id).orElseThrow();
-        postRepository.delete(post);
+
+        if (User.getCurrentUsername().equals(post.getAuthor())) {
+            postRepository.delete(post);
+        }
+        else return "redirect:/accessDenied";
 
         return "redirect:/blog";
     }
-
+    
 }
